- separating the presets to another file - removing the testedLinksCache array because I fear it would become too bulky, and it only helps lazy programmers.
As I see it now, the clickTrackingLib object will be like a glue gun, which allows you to stick click tracking functions on to dom elements. I now leave it to the end-user to keep track of which links have been changed already (and replace them if need be), which I do not think will be difficult.